Python script that queries data from Materials Project using Pymatgen API to local storafe
This script contains a python class that uses Pymatgen's MPRester to query data from the materials project website. 
The data collected include material/task_id, the chemical formula of the material, the final energy, density, band gap, energy per atom and poisson ration of the materials. 
These materials are categorized according to their respective families i.e perovskite oxides, anti-perovskite oxides, transition metal carbides and transition metal nitrides. 
More material families and properties can be added to the script depending on the intended results. 
The result of the script is highly dependent on the availability of the information on Materials Project Website